+++ lib/ruuuby/class/hsh.rb
@@ -1,0 +1,6 @@
+
+# add various functions to existing class +Hash+ (and explicitly create aliases to play nice with IDEs)
+class ::Hash
+  alias_method :🔑?, :key?
+  alias_method :∃🔑?, :key?
+end