+++ lib/ruuuby/class/nil.rb
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-# explicitly create aliases to play nice with IDEs
+# add various functions to existing class +NilClass+ (and explicitly create aliases to play nice with IDEs)
 class ::NilClass
   alias_method :empty?, :empty?
 end